+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This is a patch release with a few small bug fixes.
 
 Bug fixes:
 
+* [BUG #2654](https://github.com/BurntSushi/ripgrep/issues/2654):
+  Fix `deb` release sha256 sum file.
 * [BUG #2659](https://github.com/BurntSushi/ripgrep/issues/2659):
   Fix Fish shell completions.
 * [BUG #2662](https://github.com/BurntSushi/ripgrep/issues/2662):